+++ views/dashboard/index.php
@@ -258,10 +258,7 @@ $frequent_emails = StatisticsModule::getInstance()->frequentEmails();
                     <h5> Here are a few changes in this version:</h5>
                     <ol>
                         <!--                        <li>Pro users can now receive suggested spam words and emails every month.</li>-->
-                        <li>
-                            Update WordPress Tools
-                        </li>
-                        <li> Fix translations were triggered too early error</li>
+                        <li> Bug Fixes and improvements</li>
                     </ol>
 					<?php /*esc_html_e( "Please help translate this plugin to your language", KMCFMF_TEXT_DOMAIN ) */
 					?><!-- <a